Card actions, reducer, container, render
# There is currently a major bug where cardActions.createCard(id, name)
is called, but the dispatch is not run. This NEEDS to be resolved for
create card to work, but for the life of me I can not figure it out. May
need some help from someone with more React/JS experience. #

+++ SmoothBankCard/src/main/java/com/ss/sbank/dao/cards/CardDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ss.sbank.de.cards.Card;
 
-public interface CardDAO extends JpaRepository<Card, Integer>{
+public interface CardDAO extends JpaRepository<Card, Long>{
 
 }
